MGDOBR-103: refactoring
