+++ shard-operator/src/main/java/com/redhat/service/bridge/shard/operator/providers/TemplateProvider.java
@@ -1,0 +1,7 @@
+package com.redhat.service.bridge.shard.operator.providers;
+
+import io.fabric8.kubernetes.api.model.apps.Deployment;
+
+public interface TemplateProvider {
+    Deployment loadIngressDeploymentTemplate();
+}
